producer/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pika==1.2.0
python-dotenv
